+++ docs/changelog.rst
@@ -4,4 +4,5 @@ Changelog
 1.0.0 (unreleased)
 ------------------
 
+- #2 Fix traceback in listing when no antibiotic class was set
 - Initial Release